+++ validity/__init__.py
@@ -19,8 +19,8 @@ class NetBoxValidityConfig(PluginConfig):
     version = "3.4.3"
     base_url = "validity"
     django_apps = ["django_bootstrap5"]
-    min_version = "4.3.0"
-    max_version = "4.5.99"
+    min_version = "4.4.0"
+    max_version = "4.6.99"
     graphql_schema = "graphql.schema"
 
     # custom field
