+++ validity/netbox_changes/oldest.py
@@ -1,30 +1,7 @@
-# NetBox 4.3
+# NetBox 4.4
 from pydoc import locate
-from typing import TYPE_CHECKING
-
-from extras.events import enqueue_event, flush_events
-
-
-if TYPE_CHECKING:
-    from validity.models import ComplianceReport
-    from validity.scripts.data_models import RequestInfo
 
 
 BaseModelFilter = locate("netbox.graphql.filter_mixins.BaseFilterMixin")
 NetBoxModelFilter = locate("netbox.graphql.filter_mixins.NetBoxModelFilterMixin")
-
-
-def get_logs(job):
-    return job.data["log"]
-
-
-def set_logs(job, logs):
-    if not isinstance(job.data, dict):
-        job.data = {}
-    job.data["log"] = logs
-
-
-def enqueue(report: "ComplianceReport", request: "RequestInfo"):
-    queue = {}
-    enqueue_event(queue, report, request.get_user(), request.id, "object_created")
-    flush_events(queue.values())
+ChoicesType = locate("django.db.models.enums.ChoicesMeta")
