+++ validity/scripts/keeper.py
@@ -5,7 +5,6 @@ from core.choices import JobStatusChoices
 from core.models import Job
 
 from validity import di
-from validity.netbox_changes import set_logs
 from validity.utils.logger import Logger
 from .exceptions import AbortScript
 
@@ -48,7 +47,7 @@ class JobKeeper:
         self, status: str = JobStatusChoices.STATUS_COMPLETED, error: str | None = None, output=None
     ) -> None:
         self.job.data = self.job.data or {}
-        set_logs(self.job, [log.serialized for log in self.logger.messages])
+        self.job.log_entries = [log.serialized for log in self.logger.messages]
         output = output or self.job.data.get("output")
         self.job.data["output"] = output
         self.job.terminate(status, error)
