+++ validity/scripts/runtests/combine.py
@@ -10,16 +10,22 @@ from core.models import Job
 from dimi import Singleton
 from django.db.models import QuerySet
 from django.urls import reverse
+from extras.events import enqueue_event, flush_events
 
 from validity import di
 from validity.models import ComplianceReport, ComplianceTestResult
-from validity.netbox_changes import enqueue
 from validity.utils.logger import Logger, Message
 from ..data_models import FullRunTestsParams, RequestInfo, TestResultRatio
 from ..exceptions import AbortScript
 from ..keeper import JobKeeper
 from ..launch import Launcher
 from ..parent_jobs import JobExtractor
+
+
+def enqueue(report: "ComplianceReport", request: "RequestInfo"):
+    queue = {}
+    enqueue_event(queue, report, request, "object_created")
+    flush_events(queue.values())
 
 
 @di.dependency(scope=Singleton)
